feat: config overhaul, moved enabled to client, implement enforced
This update adds the new parameters to the config spec. Enabled is now handled client-side, allowing
clients to opt-in to certain directories. Additionally, enforced has been enabled as a way for
server owners to bypass .nosync, update files that have been changed on the client and remove all
unapproved files from clients